show latest dampening value on summary view
