+++ packages/shared/src/components/CombatReport/CombatSummary/Meters.tsx
@@ -1,6 +1,7 @@
 import _ from 'lodash';
 import moment from 'moment';
 
+import { getDampeningPercentage } from '../../../utils/dampening';
 import { Utils } from '../../../utils/utils';
 import { TimestampDisplay } from '../../common/TimestampDisplay';
 import { useCombatReportContext } from '../CombatReportContext';
@@ -37,6 +38,7 @@ export const Meters = () => {
     : 0;
   const iLvlAdvantage = friendsAvgItemLevel - enemyAvgItemLevel;
   const effectiveDuration = Utils.getEffectiveCombatDuration(combat);
+  const latestDampening = getDampeningPercentage(combat.startInfo.bracket, players, combat.endTime);
 
   return (
     <div className="flex flex-col">
@@ -63,6 +65,12 @@ export const Meters = () => {
               <td className="text-right bg-base-200">
                 {moment.utc(combat.endTime - combat.startTime).format('mm:ss')}
               </td>
+            </tr>
+            <tr>
+              <td colSpan={3} className="bg-base-200">
+                Dampening
+              </td>
+              <td className="text-right bg-base-200">{latestDampening.toFixed()}%</td>
             </tr>
             {combat.playerTeamRating ? (
               <tr>
